requirements.txt
absl-py==1.3.0
accelerate==0.21.0
aiohttp==3.8.3
aiosignal==1.2.0
albumentations==1.1.0
altair==4.2.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
blinker==1.5
blobfile==2.0.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
coloredlogs==15.0.1
comet-ml==3.31.15
comm==0.1.2
commonmark==0.9.1
configobj==5.0.6
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
datasets==2.8.0
debugpy==1.6.5
decorator==5.1.1
deepspeed==0.8.3
# diffusers @ git+git@github.com:FrankCCCCC/diffusers.git
dill==0.3.5.1
docker-pycreds==0.4.0
dulwich==0.20.46
einops==0.4.1
entrypoints==0.4
etils==0.8.0
everett==3.1.0
executing==1.2.0
filelock==3.8.0
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
ftfy==6.1.1
future==0.18.2
gdown==4.6.0
gitdb==4.0.9
GitPython==3.1.31
google-auth==2.22.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
hjson==3.1.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imageio==2.14.1
imageio-ffmpeg==0.4.7
importlib-metadata==5.0.0
importlib-resources==5.10.0
intel-openmp==2023.0.0
ipykernel==6.20.1
ipython==8.8.0
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonlines==3.1.0
jsonschema==4.16.0
jupyter_client==7.4.8
jupyter_core==5.1.3
jupyterlab-widgets==3.0.5
kaggle==1.5.12
kiwisolver==1.4.4
kornia==0.6.0
lightning-utilities==0.5.0
lpips==0.1.4
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mkl==2023.0.0
mpmath==1.3.0
multidict==6.0.2
multiprocess==0.70.13
nest-asyncio==1.5.6
networkx==3.0
ninja==1.11.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.1.1
onnx==1.14.0
onnxruntime==1.15.1
opencv-python==4.2.0.34
opencv-python-headless==4.7.0.68
packaging==21.3
pandas==1.5.2
paramiko==2.11.0
parso==0.8.3
pathtools==0.1.2
-e git+https://github.com/huggingface/peft.git@70af02a2bca5a63921790036b2c9430edf4037e2#egg=peft
peppercorn==0.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
piq==0.8.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
plumbum==1.8.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.3
ptyprocess==0.7.0
pudb==2019.2
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.7
pydeck==0.8.0
pyDeprecate==0.3.1
Pygments==2.13.0
# PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1683676063469/work
Pympler==1.0.1
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytorch-fid==0.3.0
pytorch-lightning==1.5.9
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
# pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qudida==0.0.4
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
responses==0.18.0
rich==13.0.1
rpyc==5.2.3
rsa==4.9
sacremoses==0.0.53
safetensors==0.3.1
scalablerunner @ git+https://github.com/Database-Project-2021/scalablerunner.git@1148a86c37b8e12ca3ee372ad14956acb24a881a
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.10.0
scp==0.14.4
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
semantic-version==2.10.0
semver==2.13.0
sentry-sdk==1.10.1
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.6.2
streamlit==1.16.0
sympy==1.12
tbb==2021.8.0
tensorboard==2.12.2
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow-datasets==4.7.0
tensorflow-metadata==1.10.0
termcolor==2.0.1
test-tube==0.7.5
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.10.10
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torch-fidelity==0.3.0
torchaudio==0.13.1+cu116
torchinfo==1.7.1
torchmetrics==0.11.0
torchsummary==1.5.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
transformers==4.28.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
urwid==2.1.2
validators==0.20.0
wandb==0.15.0
watchdog==2.2.1
wcwidth==0.2.5
websocket-client==1.3.3
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.14.1
wurlitzer==3.0.2
xxhash==3.1.0
yarl==1.8.1
zipp==3.10.0
